moved get s3 client method to utils
