+++ renthub/utils.py
@@ -4,11 +4,24 @@ import logging
 from enum import Enum
 
 from django.conf import settings
-from mysite.settings import get_s3_client
 import boto3
 from botocore.exceptions import ClientError
 
+from mysite.settings import AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, AWS_S3_REGION_NAME
+
 logger = logging.getLogger('renthub')
+
+
+def get_s3_client():
+    """Get S3 client for authentication to access S3 storage."""
+    return boto3.client(
+        's3',
+        aws_access_key_id=AWS_ACCESS_KEY_ID,
+        aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
+        region_name=AWS_S3_REGION_NAME
+    )
+
+
 s3_client = get_s3_client()
 BUCKET_NAME = settings.AWS_STORAGE_BUCKET_NAME
